+++ tilelang/backend/common.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 
 from tilelang.backend.pipeline import Pipeline, register_pipeline
-from tilelang.engine.pass_pipeline import CommonPassPipelineBody
+from tilelang.backend.cpu.pipeline import CPUPassPipelineBody
 
 
-for _kind in ("c", "llvm", "metal", "webgpu"):
-    register_pipeline(Pipeline(_kind, CommonPassPipelineBody))
+register_pipeline(Pipeline("webgpu", CPUPassPipelineBody))
