+++ tilelang/backend/cpu/__init__.py
@@ -1,0 +1,1 @@
+from . import pipeline  # noqa: F401
