+++ tilelang/backend/cuda/pipeline.py
@@ -5,15 +5,15 @@ from tvm.target import Target
 
 import tilelang
 from tilelang.backend.pipeline import Pipeline, register_pipeline
-from tilelang.contrib.nvcc import have_pdl, have_tma
-from tilelang.engine.pass_pipeline import (
+from tilelang.backend.pipeline_utils import (
     LayoutVisual,
     allow_vectorize,
+    should_disable_shared_memory_reuse,
     should_enable_aggressive_merge,
     should_enable_race_check,
     should_force_let_inline,
-    should_disable_shared_memory_reuse,
 )
+from tilelang.contrib.nvcc import have_pdl, have_tma
 from tilelang.transform import PassContext
 
 
@@ -59,6 +59,8 @@ def CUDAPassPipelineBody(mod: IRModule, target: Target) -> IRModule:
     mod = tilelang.transform.Simplify()(mod)
     # Set layouts for reducers
     mod = tilelang.transform.LayoutReducer()(mod)
+
+    # @CUDA-specific
     # Tile-level warp specialization: runs before layout inference so that
     # producer/consumer split happens at the high-level tile-op IR.
     # The pass classifies copy ops as TMA/cp.async/sync inline (no prior
@@ -67,17 +69,27 @@ def CUDAPassPipelineBody(mod: IRModule, target: Target) -> IRModule:
     # applies.
     if allow_warp_specialized(target=target):
         mod = tilelang.transform.ProducerConsumerWarpSpecialized()(mod)
+
+    # @CUDA / Blackwell specific
     # Lower 2SM TCGEN5MMA and related on Blackwell target (must run before
     # LayoutInference so that the use_2cta annotation is visible to infer_layout)
     mod = tilelang.transform.LowerBlackwell2SM()(mod)
+
     # Normalize if-without-else wrappers before pipeline planning. This keeps
     # pipeline body extraction focused on canonical SeqStmt bodies.
     mod = tilelang.transform.IfStmtBinding()(mod)
+
     # Run pipeline planning and software-pipeline rewriting before layout
     # inference so inferred layouts see the final pipelined structure directly.
     mod = tilelang.transform.PipelinePlanning()(mod)
     mod = tilelang.transform.InjectSoftwarePipeline()(mod)
     mod = tilelang.transform.Simplify()(mod)
+
+    # @Metal specific
+    # On Metal, rewrite local.fragment GEMM accumulators to metal.simdgroup
+    # before layout inference. simdgroup matrices are opaque and have no
+    # explicit thread-level layout, so layout inference must not see them.
+    # mod = tilelang.transform.metal.MetalFragmentToSimdgroup(mod)
 
     # Infer memory layouts for fragments and shared memory
     mod = tilelang.transform.LayoutInference()(mod)
@@ -85,6 +97,8 @@ def CUDAPassPipelineBody(mod: IRModule, target: Target) -> IRModule:
     LayoutVisual(mod)
     # Lower high-level tile operations to low-level operations
     mod = tilelang.transform.LowerTileOp()(mod)
+
+    # @CUDA specific
     # Lower l2 persistent map
     mod = tilelang.cuda.transform.LowerL2Persistent()(mod)
     # Decouple type cast vectorization constraints before vectorization
@@ -104,16 +118,22 @@ def CUDAPassPipelineBody(mod: IRModule, target: Target) -> IRModule:
     # Hoist any root-block annotations to PrimFunc attrs if pass is available
     mod = tilelang.transform.HoistNonRestrictParams()(mod)
 
+    # @CUDA-specific
     # Lower the shared.tmem into specific initialization slot
     mod = tilelang.transform.LowerSharedTmem()(mod)
-    has_tma = module_has_tma(mod)
+
     # Pipeline barriers are now created at final expanded size by
     # InjectSoftwarePipeline, so no late MVB barrier fixup is needed.
     # Buffer allocation placement is handled uniformly for both paths.
     mod = tilelang.transform.PlanAndUpdateBufferAllocationLocation()(mod)
+    # @CUDA-specific
     mod = tilelang.transform.LowerSharedBarrier()(mod)
+
+    # @CUDA-specific
+    has_tma = module_has_tma(mod)
     if has_tma:
         mod = tilelang.transform.FuseMBarrierArriveExpectTx()(mod)
+
     mod = tilelang.transform.HoistGlobalBufferAllocations()(mod)
     mod = tilelang.transform.LowerOpaqueBlock()(mod)
     mod = tilelang.transform.Simplify()(mod)
@@ -145,38 +165,55 @@ def CUDAPassPipelineBody(mod: IRModule, target: Target) -> IRModule:
     # the Legalization.
     mod = s_tir.transform.InferFragment()(mod)
     mod = tilelang.transform.LowerThreadAllreduce()(mod)
+
+    # @CUDA-specific
     mod = tilelang.transform.LowerLDGSTG()(mod)
     mod = tilelang.cuda.transform.LowerHopperIntrin()(mod)
+
     mod = tilelang.transform.AnnotateDeviceRegions()(mod)
     mod = tilelang.transform.SplitHostDevice()(mod)
+
+    # @CUDA-specific
     # Mark the function contains pdl_sync or pdl_trigger
     mod = tilelang.transform.MarkCudaSyncCalls(have_pdl(target))(mod)
     mod = tilelang.transform.AnnotateReadOnlyParams()(mod)
+
     # MergeSharedMemoryAllocations must be applied after SplitHostDevice
     # because the merged allocation site is at the beginning of each device function
     enable_aggressive_merge = should_enable_aggressive_merge(pass_ctx=pass_ctx, target=target)
     disable_reuse = should_disable_shared_memory_reuse(pass_ctx=pass_ctx)
     mod = tilelang.transform.MergeSharedMemoryAllocations(enable_aggressive_merge=enable_aggressive_merge, disable_reuse=disable_reuse)(mod)
+
+    # @CUDA-specific
     # InjectFenceProxy is a no-op on targets that lack the TMA / async-proxy
     # programming model; the pass itself checks the PrimFunc's target.
     mod = tilelang.transform.InjectFenceProxy()(mod)
+
     mod = tilelang.transform.ThreadSync("shared")(mod)
     mod = tilelang.transform.ThreadSync("shared.dyn")(mod)
+
+    # @CUDA-specific
     # Inject conservative tcgen05 fences on Blackwell (SM100+).
     # Must run after ThreadSync so that tvm_storage_sync calls are present.
     # The pass handles shared syncs and simple linear wait/use, use/arrive
     # handoffs, and is a no-op on non-SM100 targets or functions without TMEM.
     mod = tilelang.transform.InjectTcgen05Fence()(mod)
+
     mod = tilelang.transform.MergeIfStmt()(mod)
+
+    # @CUDA-specific
     # NOTE: LowerPTXAsyncCopy is applied earlier (before PipelinePlanning).
     if allow_warp_specialized(pass_ctx=pass_ctx, target=target):
         mod = tilelang.transform.AnnotateWarpGroupRegAlloc()(mod)
+
     mod = tilelang.transform.MakePackedAPI()(mod)
     mod = tilelang.transform.Simplify()(mod)
     mod = tilelang.transform.LowerDeviceKernelLaunch()(mod)
 
+    # @CUDA-specific
     # Transform threadblock to persistent threadblock
     mod = tilelang.cuda.transform.PersistThreadblock()(mod)
+
     return mod
